// src/types/PaymentProviderEvent.tsx
{/*  ===================================THIS FILE WAS AUTO GENERATED=================================== */}

export type PaymentProviderEvent = {
    paymentProviderEventID: string;
    providerName: string;
    providerEventID: string;
    eventType: string;
    processingStatus: string;
    payload: string;
    receivedAt: string;
    processedAt: string;
    failureReason: string;
    retryCount: number;
    capturedDate: string;
};